requirements.txt
numpy
scipy
gymnasium>=0.28.1
sapien==2.2.2
mplib==0.1.0
h5py
pyyaml
tqdm
GitPython
tabulate
gdown>=4.6.0
transforms3d
opencv-python
imageio
imageio[ffmpeg]
trimesh
rtree
transformers==4.32.1
sentence-transformers ==2.2.2
hydra-core==1.3.2
